Moving files into their package
